backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.3.0
click==8.3.1
colorama==0.4.6
email-validator==2.2.0
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
passlib==1.7.4
python-jose==3.5.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
SQLAlchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
